# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "service",
        "description": "Swagger OpenApi documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "get",
                "operationId": "indexActivity",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            },
            "post": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "store",
                "operationId": "storeActivity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "building_id"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/Organization/properties/name"
                                    },
                                    "building_id": {
                                        "$ref": "#/components/schemas/Organization/properties/building_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "show",
                "operationId": "showActivity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "put": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "update",
                "operationId": "updateActivity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActivityResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "delete": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "delete",
                "operationId": "deleteActivity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/append-activity/{id}": {
            "post": {
                "tags": [
                    "activities"
                ],
                "summary": "",
                "description": "append child activity",
                "operationId": "appendActivity",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AppendActivityRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "400": {
                        "$ref": "#/components/responses/404"
                    },
                    "403": {
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ActivityResource": {
                "title": "ActivityResource",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "404": {
                "description": "Не найдено",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "title": "Сообщение об ошибке",
                                    "type": "string",
                                    "example": "Сущность [Сущность][ID] не найдена"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "500": {
                "description": "Серверная ошибка",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "title": "Общее сообщение об ошибке",
                                    "type": "string",
                                    "example": "При вычислении свойств объекта произошла ошибка"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "200": {
                "description": "Успешный ответ",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "success": {
                                    "title": "Результат выполнения запроса",
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Ошибка валидации",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "title": "Общее сообщение об ошибке",
                                    "type": "string",
                                    "example": ""
                                },
                                "errors": {
                                    "title": "Массив ошибок непрошедшего валидацию поля (опциональный)",
                                    "properties": {
                                        "value": {
                                            "title": "value",
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                },
                "example": 1
            },
...
        },
        "requestBodies": {
            "AppendActivityRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "latitude",
                                "longitude",
                                "distance"
                            ],
                            "properties": {
                                "activity_id": {
                                    "type": "integer",
                                    "example": 1
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "token",
                "in": "header"
            }
        }
    },
...
}